requirements.txt
Flask>=2.0.1
Flask-Cors>=3.0.10
mysql-connector-python>=8.0.26
gunicorn>=20.1.0
requests>=2.28.1
Pillow>=5.1.0
flask-tailwind>=0.2.1
requests
google-api-python-client